Update to 1903 (10.0.18362.0) SDK. Map UINT64 to long (System.Int64). D3D12 and various improvements.

+++ src/samples/Vortice.SampleFramework/D3D12GraphicsDevice.cs
@@ -39,7 +39,7 @@ namespace Vortice
         private readonly ID3D12Resource _vertexBuffer;
 
         private readonly ID3D12Fence _d3d12Fence;
-        private ulong _fenceValue;
+        private long _fenceValue;
         private readonly AutoResetEvent _fenceEvent;
         private int _frameIndex;
 
@@ -197,7 +197,7 @@ namespace Vortice
             _vertexBuffer = _d3d12Device.CreateCommittedResource(
                 new HeapProperties(HeapType.Upload),
                 HeapFlags.None,
-                ResourceDescription.Buffer((ulong)vertexBufferSize),
+                ResourceDescription.Buffer(vertexBufferSize),
                 ResourceStates.GenericRead);
 
             var triangleVertices = new Vertex[]
